Move isna function into _vectorized
